redqueen: time limit for colorization
For nondeterministic inputs, or inputs where a large portion of the
input bytes break the path (e.g., bytes protected by a checksum),
colorization will split the range n times, once for every byte.

This is bad: for a 10k input that runs in 100ms, colorization
will take 15 minutes. For larger/slower inputs, it can easily
take hours.

This places a upper-bound on the runtime of colorization.